Pricing categories are using the wrong type
When a new pricing category is created from it's own view(i.e the view
where the form is initialised to that category) it uses the
"Imaging Procedure" pricing type rather than the correct pricing type.
So rather than just watch the category to decide when to change the
pricing type I fire it when it's about to update.