Request can be completed with nonzero Balance due
Users can complete the requests even when the patient has not payed
for the request. So I added an attribute to track payment status and
then set it when a request is about to be loaded so it can be
accessed by the button restriction code. Note that nothing was done
for procedures as they don't have completion button