Lab form doesn't indicate values not in range
The expectation of the user is that given a specific range for the
some specific lab for inputs, when the test produces a different
result the input field will indicate with some color. So I created
a special component specifically for that. Also the user expects to
be guided on what such range is. That was added to the component.