Return doesn't work for Labs/Imaging/Medication
This was because the button use the `setProperties` function which
as we saw in dealing with invoice transition doesn't work. Also
setting this property was based on the assumption that it should
not be possible to edit a completed order. Also removed the fix
to viewing lab result as requested by the TODO to that fix.