Detailed purchase report fails to show
This was cause by the discount mixin. It formats the purchase cost
after calculating hence converting it to a string, forcing others
to use a string where a number is expected. The reason it didn't
appear earlier is because "Javascript". Previously fixed the
formatting in the last commit. So I moved the formatting to the
models in this commit.